[kernel][arm64] Separate boot CPU from secondary entrypoints (1/2)
At one point it made sense to have a unified CPU start-up path for the
boot and secondary CPUs: there was a fair amount of commonality, and we
could easily switch on MPIDR_EL1 for context-specific codepaths.
However, we are now in the process of both cutting down on that
commonality and the logic in the boot CPU path altogether, so it makes
sense now to separate them.

This is the first in a series of two changes that accomplishes this,
separated out for easier review. The first patchset creates secondary.S
with a plain `cd zircon/kernel/arch/arm64; cp start.S secondary.S`. The
second patchset does a _minimal_ removal of content in secondary.S to
ensure that things work. Proper clean-up will happen in the second
change of the series and reviewers of this one should look at
https://fuchsia-review.googlesource.com/c/fuchsia/+/1204364/1..2

Bug: 42164859
Change-Id: Ie1d5593a8dc98398a2175f8ea8e9dbbd071e6b94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204364
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>