[kernel][vm] Augmend mapping WAVL tree
Used an augmented WAVL tree to store mappings in a VmObject instead of
list. This allows for efficiently finding the subset of mappings that
might overlap with an invalidation range, instead of having to always
process the entire list.

When there are many non-overlapping mappings this dramatically reduces
the number of mappings visited, and greatly improves performance.

Bug: 310255065
Change-Id: Ib39b4ff1c386fd25cff716fee04a31d34cbba12f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1144972
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>