[fxfs] Reduce memory usage of the DirentCache
Don't initialize the HashMap backing the cache to the DirentCache's
limit. Each volume has a DirentCache but only the blob volume gets close
to the limit. Reserving space for 2000 entries for the unencrypted
volume which only has 2 files is a waste. Growing the HashMap as
necessary doesn't significantly affect performance.

Shrink the HashMap backing the cache when shrinking the cache size. The
number of items in the cache is reduced when under memory pressure but
the HashMap backing the cache still had capacity for the larger number
of items. The HashMap is now shrunk to further reduce memory usage when
under memory pressure.

The cache size when under normal memory pressure has been reduced to
2000. The blob volume stays just under this limit at boot.

Bug: 393365596
Change-Id: Ic126cc1c6c5217b32b8877d53fca93db7d33994e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1203845
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>