[gidl][fuchsia-controller] Add OWNERS for fuchsia-controller backend
Both awdavies@ and chcl@ are building fidlgen_python and will be making
changes to the fuchsia-controller GIDL backend as we discovers parts of
the backend that should change. For example, these is a follow-up change
to avoid using the Python keyword `bytes` as a variable in tests.

Change-Id: I13a866c7fee6b9319b01e9ebd57deef8e2d20f1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1202424
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>