[bazel][ffx] Add SDK manifest to ffx product create
This adds the explicit configuration to use the SDK manifest when
running `ffx-product create`. This is needed since the SDK version is
read from the SDK when building the product bundle.

This change is in preparation for not implicitly reading the
SDK manifest in-tree, which will be done in a follow-on CL.

Change-Id: Iee9246270af83f77aec9dc29f1acb523bb16eddc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1209104
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>