[tee] Implement TEE_IsAlgorithmSupported
We implement TEE_IsAlgorithmSupported() to indicate support for the
current (limited, and growing) set of supported algorithms, replacing
a previous panic. We will naturally update its implementation as we
support more algorithms.

This makes a number of WTPI tests that were previously passing start to
fail, as they had been indirectly relying on the call to panic, which is
something that it should never do. This also makes a handful of xtest
cases begin to pass by virtue of the test case gracefully skipping when
encountering an unsupported algorithm.

Bug: 360942581
Change-Id: Ifec5b2da8edb1a94a040cf45614c1b02e5a10a10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1199824
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Robinson <jamesr@google.com>