[netstack] Encode skips in ANVL test expectations
Encode test cases that should be skipped or that will be skipped by
ANVL in expectations as Skip and AnvlSkip respectively.

Delete the test expectations for DHCPv6 client PD against NS2 because
the suite is only run against NS3 due to license contention in Infra,
and the presence of expectations now implies that a suite is run.

A follow-up change in the network-conformance repo will omit the
skipped test cases when running the full suites instead of encoding
the skips in GN.

Bug: 355222753
Change-Id: I9444a121b5614cd999412927f2f999923702aee5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204745
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Tony Gong <gongt@google.com>