[starnix] Test runner removes test_type from start info.
This allows the starnix kernel to be stricter about the program
blocks it runs by removing a key which is not known to the kernel.

By being stricter when parsing the program block we can provide
better errors on typos.

Bug: 308055654
Change-Id: I02b2599ceff98c62b33d87a78965ca8f61a12337
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1208739
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>