[runtime] Optimize extra atomic acquire load in _zx_thread_self
libsync and pthread mutexes call _zx_thread_self() when acquiring a lock
to retrieve a 32 bit thread identity value with the low bit clear. This
function retrieves the handle value at an offset from the thread
pointer using zxr_thread_get_handle() to do the lookup. This function
can also be used to retrieve the handle value for a different thread and
as such a synchronizing atomic acquire load was introduced in commit
1db2dc84f6f24fee118f750e8ea8aa2278d9e5de to make these cases safe. This
load is not needed when retrieving the calling thread's handle value.

This adds a zxr routine to be used only for the calling thread's value
and uses that in _zx_thread_self(). This removes an atomic acquire load
from the mutex acquisition path.

Change-Id: Icc1fc87874ecf1f9bc2b0b5cc3945abbcea3ab8f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207605
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>