[fshost] potentially set a type guid for test disk
The gpt matcher will match anything that doesn't have a type guid, or
anything where the type guid is "empty" (all zeros), which is what the
v2 ramdisk reports when it's not configured with a guid. It's enabled by
default so it ends up grabbing any devices that fall through the cracks.

This adds a default type guid that is non-zero, and the ability to
configure it. If the device is uninitialized or gpt is set, the guid is
cleared so it'll trigger the matcher.

At some point we will make the gpt matcher configured via args instead
of heuristically picking a device, which will remove the need for this.

Bug: 391891190
Change-Id: I76c564a3bb37bcfeb4658eb4175076e2b5ac2206
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1209967
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>