[kernel][vm] Thread blocked in usercopy can be killed
Fixes a bug where a thread blocked on a page request from a usercopy
could end up busy looping on receiving a kill signal. Since we do not
return out to user mode, the kill signal cannot be processed in the same
way as a user mode page fault, where we pretend that the page fault
succeeded.

This was a problem only for usercopy with inline fault resolution. Added
a test that does not terminate without this fix. Also added a test for
the usercopy variant that captures faults (this worked fine previously).

Test: core-pager-process
Run-All-Tests: true
Bug: 42084841
Change-Id: Ie849d6441881cbcafd69de72827160abb0fd16b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1205447
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>