[pigweed] Required changes to support protobuf 28.0
In order to roll protobuf 28.0 into the Fuchsia tree, the pigweed
specific pw_protobuf_compiler backend build rules must be adjusted
to reflect the fact that the python sources from
//third_party/protobuf/python are no longer directly usable.

The issue is the python_edition_defaults.py file which must be now
generated at build time.

This CL adds the necessary operations, and a new addition to the list
of required python sources to the protoc python package.

Note that this CL requires https://fxrev.dev/1188392 in order to
update //build/secondary/third_party/protobuf, as well as
tqr/930471 to roll both protobuf and pigweed at the same time,
otherwise the build will break.

Bug: 386798036
Change-Id: Ie6a71cd25a5b3797f2dd0085c026c79c1fb8b354
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1191112
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Ali Saeed <saeedali@google.com>