[testing] Sanitize child component names in enumeration_test_adapter
It is natural to give the name of the case in the enumeration config the
same name as the canonical test case name, so we do that. Further. gtest
will generate a parameterized test case name with slashes (like
HmacTests/HmacTest.HMAC_SHA256Basic/0). Currently however - also out of
want of a canonical name - the enumeration adapter forwards the name of
the test case as the name of the child component in which it's run - and
slashes are invalid in component names.

Accordingly, we sanitize the constructed component names, replacing
slashes with, say, underscores.

This changes starts making an number of tests start to actually run
instead of short-circuiting and erroring out. In the case of the
oemcrypto-integration tests the runtime is now no longer reasonable in
its current form, so we disable most of its cases for now (pending a
closer look out-of-band from this change). In the case of the HMAC
tests, these generate UBSan failures and more investigation is needed.

Bug: 397716819
Change-Id: I0b54fe695d89df4941933d29b98f8d50d1408d12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210568
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: James Robinson <jamesr@google.com>