[kernel][vm] Log pager creating process name
Include the name of the process that created the pager in pager delay
and timeout logs.

Adds an fbl::Name<ZX_MAX_NAME_LEN> to the PagerDispatcher object, but
there should not be many pagers getting created in practice, so this
seems like a fine tradeoff for better debuggability.

Bug: 393202368
Change-Id: I0bcec0893ffc06182558c993195922309a656e8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210048
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>