[nanohub][starnix] Add nanohub_socket_file for sysfs EOF
This is intended to be a stop-gap solution to allow us to use sockets
for sysfs endpoints with predictable behavior. The consumers of our
interface expect to receive an EOF (or syscall or read returning 0)
after their first read containing the full payload. We are able to send
the responses in a single datagram, so this solution would carry us
while a more robust interface is developed.

Adds additional logic to underlying test to ensure EOF behaves as
expected.

Test: starnix_nanohub_module_test
Bug: 391691908
Change-Id: If559c255152cfdb9de59407598e3b8b6de3e525e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200726
Reviewed-by: Eric Rahm <erahm@google.com>
Commit-Queue: Duncan Millard <duncanmillard@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>