[system-update-committer][cleanup] Remove hardcoded Config
Currently, this config is hardcoded to either require
all health checks to pass or to ignore all health checks.

Health checks are already ignored if in the recovery configuration
so it is redundant to add a config to ignore health checks for
recovery-fdr.

This CL removes the config while still following the same logic
paths. Many tests are removed because they tested configuration
values that can never occur in production.

This is part of the work to improve health verification.

Run-All-Tests: true

Fixed: 397222040

Change-Id: Ia91627fe781bba65e1b52c51e37d96f3a432c911
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1208704
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Olivia Nordquist <nolivia@google.com>