[ffx] Add trait for ffx test execution.
This adds a FfxExecutor trait that allows for commands that are not just
Isolate to run a list of commands in sequence and evaluate their success
cases.

Currently the libraries are not used outside of the emulator integration
test code, which is not in any known build paths.

Change-Id: I852a02e996a3901ac9b90303be28ec0f8aab3a5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210308
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>