[storage] BlobWriter prints checksum on corruption error
To help determine where the corruption is occurring, when BytesReady
fails with IO_DATA_INTEGRITY BlobWriter now prints the total number of
bytes written to the VMO so far and the crc32 checksum of those bytes,
which will appear in the pkg-resolver's syslog.

This should not affect production devices because BlobWriter is not
used in those configurations.

Bench tested that the printed checksum matches that of the crc32 CLI
tool.

Bug: 378158208
Change-Id: I373af6a090740522d3b3a9d8db507315086179ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211134
Commit-Queue: Ben Keller <galbanum@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>