[ffx] emu stop integration test.
Refactors the existing emulator integration tests to work and to be
included in the existing source base. For the time being, this keeps the
non-strict use-cases out of testing, as their stability is not known.

This doesn't cover the full gamut of `ffx --strict ... emu stop` tests,
as that depends on `ffx --strict ... emu start` functionality as well.

This change also moves `ffx emu *` out of being included in the main ffx
binary into a standalone tool by default.

Bug: 392655729
Fixed: 395975823

Change-Id: Icaf0449692e43b84b092f5030ad6e9162d644716
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1204879
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>