[fidl][ffx] Apply category markers to all FIDL sub-targets
Specifically, apply it to the FIDL library on the FIDL toolchain, which
all bindings types will inherit from `fidl_ir_target`. This catches uses
of FIDL libraries within dependencies of ffx rather than just use of
them via the Rust bindings. It also makes it possible to use the markers
in contexts other than ffx.

Update the allowlist to include additional dependencies caught by the
now more thorough checks.

Note: This could catch uses of FIDL libraries or types internal to a
component such that they do not pose a compatibility concern. However,
this should be rare and can be handled if/when it is a problem.

Bug: 314822328
Change-Id: I3e1ca8f60e0fd03ac6d90316bf5bdfcb6da3ac01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1207239
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>