[lacewing][ffx]: Support --net user emulators
This change allows the use of --net user emulators by setting the local
mobly driver to always use the device ip instead of using the target
name. This will mean that users will have to be aware of rebooting the
device or any change to the device ip. We have a warning in place and
will revisit if this becomes an issue.

Fixed: b/394940615
Fixed: b/366067471

Change-Id: I0d7133657815297a9d7196a121c73be5fad9f3fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206705
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Patrick Lu <patricklu@google.com>