[fidl][build] Clarify check for internal zx library
When a target name does not start with a colon, GN's
`get_label_info()` adds an extra directory to the result. In this case,
that real directory exists and is a different library. Specify a colon
so that the literal string reflects the actual library location.

Bug: 325669391
Change-Id: I42d748bcdbaa9289267b544993da2f027901f752
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211888
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>