[assembly] Add amber-files back to multi-product builds
As is turns out... we had fyi builders that run tests that need
universe packages. I am adding this back with if-conditionals
to avoid including packages from the product bundles. This is because
in a multi-product build, we may not have a product bundle set. And
if we do... that "main" product bundle is only one of the many
possible product bundles. Best to not include any product bundle
until we understand more.

Bug: 398295610
Change-Id: Ib83a9624c8aa0edc928bbbfc1cf0e2730c294a24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1212564
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>