[developer] Ensure /svc is opened with PERM_READABLE
This is required so that service instances can be enumerated correctly
when we transition fdio_service_connect to use the new Open method,
which enforces the new set of fuchsia.io rights. It currently uses the
DeprecatedOpen method which does not enforce these rights.

Bug: 324111518
Bug: 376575307
Change-Id: Ifa67216aecf4e3e30eb58860b29f1ec2c8679002
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211865
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>