[ffx][sdk] Allow specifying SDK manifest path
This change refactors SdkRoot to allow specifying the relative
path to the SDK manifest within the SDK root. Previously, it was
hardcoded to be 1 of 2 possible paths.

The previous behavior is maintained if the manifest path is None,
so we can adopt being explicit about the in-tree manifest gradually.

Bug: 397989792
Change-Id: Ib7bbf6605ad5eff589c6852434f004c1ac2d043c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211164
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>