[kernel][vm] Split RangeChangeUpdateLocked
Splits the existing range change update into two separate methods, one
that just operates on the current node, and one that just operates on
the child nodes. When operating on the self node the caller is
additionally required to indicate whether they intend to invalidate the
children later, or if they assume that they have no children.

Performing this split is a no-op in this commit and later commits in
the fine grained locking transition will take advantage of it.

Bug: 338300943
Change-Id: Ide9425b4fe895aa5e08a95218c44a7d932b15f2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1209206
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>