[unwinder] Clarify CFI loading code
This change clarifies the usages of both which style of CFI we're
reading. The eh_frame section contains "exception frames" which are
comprised of CFI records very similar, but not the same as, CFI records
contained in the debug_frame section. Within each of these is the
possibility that the CFI records are in 64 or 32 bit format, depending
on various compiler options.

It should now be more obvious to readers at a glance what kind of CFI
format we're reading from a given module. There should be no observable
changes to behavior.

Bug: https://fxbug.dev/292107156
Change-Id: I704c601949ff211e10e861f92e1d002ab25a5edf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1195603
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>