[block_server] Propagate block flags from device
As part of this, I made the `name` field in PartitionInfo non-optional,
as the only device which behaves like a partition but doesn't set is the
ramdisk.  I had to slightly modify FxblobMatcher to accommodate this, as
now the label will come back as an empty string rather than an error.

We might need to eventually allow the C block_server bindings to create
DeviceInfo::Block, since some future block devices that we will port to
using block_server might only be blocks and not also partitions.

Fixed: 376745136
Change-Id: I61258aca9527930e66a424a99efd2d7fcdb57704
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1213148
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>