[fidl][rust] Include generation in transaction ids
This adds a 7 bit "generation" in upper bits of the transaction id. It
increments each time we allocate a new txid for a channel. We mask it
off when we convert it to a slab index. This means that subsequent FIDL
calls won't end up with the same txid, even if they're in the same slab
spot.

This means that in the common pattern of a channel being used for
repeated hanging-gets each of the gets will have a distinct txid, which
allows for easier tracing.

Change-Id: I9a6acdcd6a17c0ea1131bd4f74eff26889086408
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1200091
Reviewed-by: David Koloski <dkoloski@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>