[kernel][vm] Simplify range change reasoning
Moves the range change update in ZeroPagesLocked out of the if into
the main body of the method. Guarding the range change by the if is
correct, since if can_decommit_zero_pages is false we will not be able
to remove any pages or otherwise modify the page list in a way that
would require a range change update, but this correctness is subtle.

Given that performing a zero range on a contiguous vmo is a very
uncommon operation, performing an additional unmap in this case is
worth the additional clarity.

Change-Id: I408bb8ac68fee419cd498bb939760b9f1514d613
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1213765
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>