[flash-kola] Only limit to one fastboot device if ANDROID_SERIAL unset.
This check is a useful guardrail for cases where it's not clear the
user knows which device they are flashing.

This environment variable tells fastboot which target to use.
In the absence of proper argument parsing this is the best indication
we have that a caller of this script knows which fastboot device
they'd like to flash.

Change-Id: I822fdec29f5d63c2b987c0720c82d10d162c1059
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1214090
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>