[display][coordinator] Make IdMappable ID immutable and private.
We hope to remove IdMappable in the long term. However, for the
short-to-medium term, we'll probably have to live with it.

This CL addresses the biggest issues in IdMappable -- the data member
that serves as the map key is both public and mutable.

1. Public data members go against the C++ style guide.

2. Most map implementations (including the ones in fbl) assume that an
   object's key remains immutable while the object is inserted in the
   map. Making the map key immutable is a straightforward way to ensure
   that the assumption above is true.

Having an immutable data member makes the IdMappable class non-copyable
and non-movable. The C++ style guide requires stating these aspects
explicitly.

Bug: 336394440
Test: fx test //src/graphics/display
Change-Id: I8209919a6481373515179dfcf8f876d2bb94731e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211122
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Reviewed-by: Yilong Li <liyl@google.com>