[ffx][e2e] Starting emulator synchronously
This changes the e2e_emu lib to run `ffx emu start` in-line vs. as a
background child process. The advantage to running it like this is that
the command does not return until the emulator instance is up and can be
connected to via RCS. When run as a Child process, it returns quickly,
potentially before the emulator is ready.

Bug: 398962542
Change-Id: I73de65b51804ac30ec6cda58af4b44334bd965f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1214185
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>