[ld] Let PT_INTERP string determine fuchsia.ldsvc.Loader/Config message
This makes the startup dynamic linker look at the executable's
PT_INTERP string (which presumably was used to find the dynamic
linker's own binary now running).  If it is the canonical string
expected, but has a `.../` prefix, then `...` is taken as the
string to send in the Config() FIDL message to the loader service
that directs it to look in alternate places for DT_NEEDED files.

This also cleans up all the packaging and test logic related to
libprefix in accordance with this scheme.  The startup dynamic
linker is packaged using the libprefix of dependent executables
without regard to the variant where ld-startup itself was built.
The test packaging now uses the libprefix under the per-test lib/
subdirectories so they more closely resemble libprefix use in
normal Fuchsia packages.  The test fixture logic now looks up the
PT_INTERP in each test_executable() to drive its expectations
about Config() messages as well as where to find the files to
prime the mock for expected LoadObject() requests.  In the
special cases where there is no matching PT_INTERP to use, the
tests use the TestElfLoadSet machinery to find libprefix records
stored at build time.

The ld-stub packaging is made canonical and never uses libprefix,
as it's really just data not accessed via the loader service.
Instead, it goes into lib/ but with a name that represents the
ELF layout and machine so that one remote dynamic linker package
can serve disparate ld::RemoteDynamicLinker instantiations.

Fixed: 42084623
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: I39d611dbce97991eb302ea48d563cde8f2493fab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1211524
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>