[setui][healthchecks] Add health check server to setui
Updates the cml to expose the health check to the framework
and adds the fidl service to the apropos service directory..

If the setui environment is able to be created, then we know
that all of the agents have responded without error. If the
environment fails, then setui reboots. An inability to
get a response from the component for its health check status
is treated as a failure. So, if the health check protocol is
able to be queried, we have enough information to deem the
component healthy :)

Tested: fx test //src/settings

Bug: 397979575

Change-Id: Ib8537ded06c6616b50bf88d04650f5fec60e2530
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210344
Reviewed-by: Paul Faria <paulfaria@google.com>
Commit-Queue: Olivia Nordquist <nolivia@google.com>