[kernel][vm] Dump pager and port handles on timeout
Dump all processes that have handles to the pager object and the
associated port object on a pager timeout.

This walks the root job tree recursively, but the system is in a bad
state anyway, so it should be okay to be inefficient here for better
debuggability.

Bug: 393202368
Change-Id: I0fb7b2d0f86bdffa059f67b2d7b6f193bae2d426
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1214095
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>