[ulib][concurrent] Create a vDSO compatible version of libconcurrent.
Add a new target to `ulib/concurrent` called `concurrent_vdso` which
is suitable for inclusion in a vDSO image.

`libconcurrent` makes us of a _lot_ on std::atomics.  This actually
can be an issue when building for ARM64 since some toolchains these
days automatically attempt to "outline" calls to the atomics.  ARM
v8.1 added "LSE" atomics to the instruction set, which can provide
performance advantages, but need to be tested for at runtime to
maintain full portability.

So, compilers sometimes add a call (executed either during startup, or
lazily) which will check the processor to see if it supports LSE
atomics.  Calls to atomics are not inlined, but instead go through a
thunk which can be controlled by the result of this test.
Unfortunately, none of this works in a vDSO environment.  If we wanted
to add dynamic support for LSE atomics, the decision to use them (or
not) would need to be done when we create the vDSO image, not
dynamically.

So, on ARM64, we want to pass `-mno-outline-atomics` to suppress this
behavior when we build for a vDSO.

See also b/398297232

Change-Id: I5afd911b1c2add4e28c8e7002ba1f1adc1b63c37
Bug: 398297232
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1209085
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>