[starnix][functionfs] Separate reads and writes
Reads must not be reordered relative to other reads, and writes must not
be reordered relative to other writes.

Because reads and writes were handled on the same thread, but using a
concurrent dispatcher, they may execute out-of-order relative to
other commands of the same type.

This change updates the logic to allow reads and writes to execute
independently of each other, but keeps the ordering consistent within
each type.

Bug: 395635052

Change-Id: Ic74927b16d769332e4d3a6498ba2ae17c0d49d83
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1214444
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: David Song <wintermelons@google.com>